+++ ZeroLibraries/Common/Math/GenericVector.inl
@@ -12,162 +12,162 @@
 namespace Math
 {
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension>::GenericVector()
 {
   ZeroOut();
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension>::GenericVector(scalar* data)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] = data[i];
 }
 
-template <typename ScalarType, uint Dimension>
-typename GenericVector<ScalarType,Dimension>::scalar& GenericVector<ScalarType,Dimension>::operator[](uint index)
+template <typename ScalarType, size_t Dimension>
+typename GenericVector<ScalarType,Dimension>::scalar& GenericVector<ScalarType,Dimension>::operator[](size_t index)
 {
   ErrorIf(index > Dim, "Math::GenericVector - Subscript out of range.");
   return array[index];
 }
 
-template <typename ScalarType, uint Dimension>
-typename GenericVector<ScalarType,Dimension>::scalar GenericVector<ScalarType,Dimension>::operator[](uint index) const
+template <typename ScalarType, size_t Dimension>
+typename GenericVector<ScalarType,Dimension>::scalar GenericVector<ScalarType,Dimension>::operator[](size_t index) const
 {
   ErrorIf(index > Dim, "Math::GenericVector - Subscript out of range.");
   return array[index];
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension> GenericVector<ScalarType,Dimension>::operator-(void) const
 {
   SelfType ret;
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret[i] = -array[i];
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::operator*=(scalar rhs)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] *= rhs;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::operator/=(scalar rhs)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] /= rhs;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension> GenericVector<ScalarType,Dimension>::operator*(scalar rhs) const
 {
   SelfType ret;
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret[i] = array[i] * rhs;
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension> GenericVector<ScalarType,Dimension>::operator/(scalar rhs) const
 {
   SelfType ret;
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret[i] = array[i] / rhs;
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::operator+=(SelfParam rhs)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] += rhs[i];
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::operator-=(SelfParam rhs)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] -= rhs[i];
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension> GenericVector<ScalarType,Dimension>::operator+(SelfParam rhs) const
 {
   SelfType ret;
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret[i] = array[i] + rhs[i];
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension> GenericVector<ScalarType,Dimension>::operator-(SelfParam rhs) const
 {
   SelfType ret;
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret[i] = array[i] - rhs[i];
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension> GenericVector<ScalarType,Dimension>::operator*(SelfParam rhs) const
 {
   SelfType ret;
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret[i] = array[i] * rhs[i];
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 GenericVector<ScalarType,Dimension> GenericVector<ScalarType,Dimension>::operator/(SelfParam rhs) const
 {
   SelfType ret;
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret[i] = array[i] / rhs[i];
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::operator*=(SelfParam rhs)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] *= rhs[i];
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::operator/=(SelfParam rhs)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] /= rhs[i];
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::ZeroOut()
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] = scalar(0);
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 void GenericVector<ScalarType,Dimension>::Splat(scalar value)
 {
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     array[i] = value;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 typename GenericVector<ScalarType,Dimension>::scalar GenericVector<ScalarType,Dimension>::Dot(SelfParam rhs) const
 {
   scalar ret = scalar(0);
-  for(uint i = 0; i < Dim; ++i)
+  for(size_t i = 0; i < Dim; ++i)
     ret += array[i] * rhs[i];
   return ret;
 }
 
-template <typename ScalarType, uint Dimension>
+template <typename ScalarType, size_t Dimension>
 typename GenericVector<ScalarType,Dimension>::scalar GenericVector<ScalarType,Dimension>::LengthSq(void) const
 {
   return Dot(*this);
